Add NC recurrence history API and render in admin cards
